multiple testability improvements
- separated integration tests for docker environments
- use templates for svgs
- golangci-lint compatibility (errcheck, deadline, so much more)
