+++ badge.go
@@ -3,32 +3,34 @@ package main
 import (
 	"bytes"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"math"
 	"strconv"
 	"strings"
 	"text/template"
-
-	"github.com/go-redis/cache"
-)
-
-const (
-	// Badge templates to generate badges
-	curveBadge = `<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="{{.Width}}" height="20"><linearGradient id="b" x2="0" y2="100%"><stop offset="0" stop-color="#bbb" stop-opacity=".1"/><stop offset="1" stop-opacity=".1"/></linearGradient><clipPath id="a"><rect width="{{.Width}}" height="20" rx="3" fill="#fff"/></clipPath><g clip-path="url(#a)"><path fill="#555" d="M0 0h61v20H0z"/><path fill="{{.Color}}" d="M61 0h53v20H61z"/><path fill="url(#b)" d="M0 0h114v20H0z"/></g><g fill="#fff" text-anchor="middle" font-family="DejaVu Sans,Verdana,Geneva,sans-serif" font-size="110"><text x="315" y="150" fill="#010101" fill-opacity=".3" transform="scale(.1)" textLength="510">{{.Label}}</text><text x="315" y="140" transform="scale(.1)" textLength="510">{{.Label}}</text><text x="{{.StatusX}}" y="150" fill="#010101" fill-opacity=".3" transform="scale(.1)" textLength="">{{.Status}}</text><text x="{{.StatusX}}" y="140" transform="scale(.1)" textLength="">{{.Status}}</text></g> </svg>`
-	flatBadge  = `<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="{{.Width}}" height="20"><g shape-rendering="crispEdges"><path fill="#555" d="M0 0h61v20H0z"/><path fill="{{.Color}}" d="M61 0h57v20H61z"/></g><g fill="#fff" text-anchor="middle" font-family="DejaVu Sans,Verdana,Geneva,sans-serif" font-size="110"><text x="315" y="140" transform="scale(.1)" textLength="510">{{.Label}}</text><text x="{{.StatusX}}" y="140" transform="scale(.1)" textLength="">{{.Status}}</text></g> </svg>`
 )
 
 var (
-	curveBadgeTmpl *template.Template
-	flatBadgeTmpl  *template.Template
+	curvedBadgeTmpl *template.Template
+	flatBadgeTmpl   *template.Template
 )
 
 func init() {
+	curvedBadge, err := ioutil.ReadFile("./templates/badges/curved.tmpl")
+	if err != nil {
+		panic(fmt.Sprintf("curved badge template not found: %v", err))
+	}
+
+	flatBadge, err := ioutil.ReadFile("./templates/badges/flat.tmpl")
+	if err != nil {
+		panic(fmt.Sprintf("flat badge template not found: %v", err))
+	}
+
 	tpl := template.New("")
-	curveBadgeTmpl, _ = tpl.Parse(curveBadge)
+	curvedBadgeTmpl, _ = tpl.Parse(string(curvedBadge))
 	tpl2 := template.New("")
-	flatBadgeTmpl, _ = tpl2.Parse(flatBadge)
+	flatBadgeTmpl, _ = tpl2.Parse(string(flatBadge))
+
 }
 
 type badge struct {
@@ -37,39 +39,6 @@ type badge struct {
 	Label   string
 	Color   string
 	Width   int
-}
-
-// getBadge gets the badge from img.shields.io and return as []byte
-func getBadgeImgShield(color, style, percent string) string {
-	cacheKey := fmt.Sprintf("%s-%s-%s", color, style, percent)
-	str := ""
-	err := redisCodec.Get(cacheKey, &str)
-	if err == nil {
-		return str
-	}
-	errLogger.Println(err)
-
-	imgURL := fmt.Sprintf("https://img.shields.io/badge/cover.run-%s25-%s.svg?style=%s", percent, color, style)
-
-	resp, err := httpClient.Get(imgURL)
-	if err != nil {
-		errLogger.Println(err)
-		return ""
-	}
-	defer resp.Body.Close()
-	svg, err := ioutil.ReadAll(io.LimitReader(resp.Body, 1024))
-	if err != nil {
-		errLogger.Println(err)
-	}
-	err = redisCodec.Set(&cache.Item{
-		Key:        cacheKey,
-		Object:     string(svg),
-		Expiration: -1,
-	})
-	if err != nil {
-		errLogger.Println(err)
-	}
-	return string(svg)
 }
 
 // getBadge is a function which will generate SVG rather than fetch from img.shield.io
@@ -127,7 +96,7 @@ func getBadge(color, style, status string) string {
 
 	switch style {
 	case "flat", "curve", "flat-curve":
-		_ = curveBadgeTmpl.Execute(buf, b)
+		_ = curvedBadgeTmpl.Execute(buf, b)
 	default:
 		_ = flatBadgeTmpl.Execute(buf, b)
 	}
@@ -136,15 +105,13 @@ func getBadge(color, style, status string) string {
 }
 
 // coverageBadge returns the SVG badge after computing the coverage
-func coverageBadge(repo, tag, style string) (string, error) {
+func coverageBadge(repo, tag, style string) string {
 	obj, err := repoCover(repo, tag)
 	if err != nil {
 		if err == ErrQueued {
-			return getBadge("lightgrey", style, "queued"), nil
-		}
-
-		if err == ErrCovInPrgrs {
-			return getBadge("yellowgreen", style, "testing"), nil
+			return getBadge("lightgrey", style, "queued")
+		} else if err == ErrCovInPrgrs {
+			return getBadge("yellowgreen", style, "testing")
 		}
 
 		errLogger.Println(err)
@@ -154,7 +121,7 @@ func coverageBadge(repo, tag, style string) (string, error) {
 
 	percent, err := strconv.ParseFloat(strings.Replace(obj.Cover, "%", "", -1), 64)
 	if err != nil {
-		return getBadge(color, style, "error"), nil
+		return getBadge(color, style, "error")
 	}
 
 	if percent >= 70 {
@@ -164,5 +131,5 @@ func coverageBadge(repo, tag, style string) (string, error) {
 	}
 
 	status := strconv.FormatFloat(math.Round(percent*100)/100, 'f', -1, 64)
-	return getBadge(color, style, status+"%"), nil
+	return getBadge(color, style, status+"%")
 }
