+++ main.go
@@ -48,11 +48,6 @@ const (
 	// DefaultTag is the Go version to run the tests with when no version
 	// is specified
 	DefaultTag = "1.10"
-	// cacheExpiry is the duration in which the cache will be expired
-	cacheExpiry = time.Hour
-	// refreshWindows is the time duration, in which if the cache is about to expire
-	// cover run is started again.
-	refreshWindow = time.Minute * 10
 
 	// Redis Errors
 	redisErrNil      = "redis: nil"
@@ -66,7 +61,7 @@ var (
 	// qLock is used to push to Redis channel because redis pub-sub in go-redis is
 	// not concurrency safe
 	qLock = sync.Mutex{}
-	// qChan is used to control the number of simultaneos executions
+	// qChan is used to control the number of simultaneous executions
 	qChan = make(chan struct{}, coverQMax)
 
 	httpClient = &http.Client{
@@ -260,7 +255,9 @@ func computeCoverage(stdOut string) string {
 // - Before starting evaluation, it sets the repo's status as in progress
 // - Removes the inprogress status of a repo after it's done
 func cover(repo, langVersion string) error {
-	setInProgress(repo, langVersion)
+	if err := setInProgress(repo, langVersion); err != nil {
+		return err
+	}
 
 	stdOut, stdErr, err := run(langVersion, repo)
 	if err != nil {
@@ -270,7 +267,9 @@ func cover(repo, langVersion string) error {
 		}
 	}
 
-	unsetInProgress(repo, langVersion)
+	if err := unsetInProgress(repo, langVersion); err != nil {
+		return err
+	}
 
 	obj := &Object{
 		Repo:   repo,
@@ -353,28 +352,6 @@ type Repository struct {
 	Cover string
 }
 
-func repoLatest() ([]*Repository, error) {
-	repos := make([]*Repository, 0)
-	keys, _, err := redisRing.Scan(0, "*", 10).Result()
-	if err != nil {
-		errLogger.Println(err)
-		return repos, err
-	}
-
-	var obj Object
-	for _, key := range keys {
-		if len(repos) == 5 {
-			return repos, nil
-		}
-		if err := redisCodec.Get(key, &obj); err == nil {
-			if obj.Output {
-				repos = append(repos, &Repository{obj.Repo, obj.Tag, obj.Cover})
-			}
-		}
-	}
-	return repos, nil
-}
-
 // subscribe subscribes to the Redis channel
 func subscribe(qname string) {
 	pubsub := redisClient.Subscribe(qname)
@@ -387,7 +364,9 @@ func subscribe(qname string) {
 		}
 		repo, tag := repoTagFromFullName(msg.Payload)
 		qChan <- struct{}{}
-		go cover(repo, tag)
+		go func() {
+			_ = cover(repo, tag)
+		}()
 	}
 }
