+++ templates/badges/curved.tmpl
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="{{.Width}}" height="20">
+  <linearGradient id="b" x2="0" y2="100%">
+    <stop offset="0" stop-color="#bbb" stop-opacity=".1" />
+    <stop offset="1" stop-opacity=".1" />
+  </linearGradient>
+  <clipPath id="a">
+    <rect width="{{.Width}}" height="20" rx="3" fill="#fff" />
+  </clipPath>
+  <g clip-path="url(#a)">
+    <path fill="#555" d="M0 0h61v20H0z" />
+    <path fill="{{.Color}}" d="M61 0h53v20H61z" />
+    <path fill="url(#b)" d="M0 0h114v20H0z" />
+  </g>
+  <g fill="#fff" text-anchor="middle" font-family="DejaVu Sans,Verdana,Geneva,sans-serif" font-size="110">
+    <text x="315" y="150" fill="#010101" fill-opacity=".3" transform="scale(.1)" textLength="510">{{.Label}}</text>
+    <text x="315" y="140" transform="scale(.1)" textLength="510">{{.Label}}</text>
+    <text x="{{.StatusX}}" y="150" fill="#010101" fill-opacity=".3" transform="scale(.1)" textLength="">{{.Status}}</text>
+    <text x="{{.StatusX}}" y="140" transform="scale(.1)" textLength="">{{.Status}}</text>
+  </g>
+</svg>
